pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-exercises"
version = "0.1.0"
description = "A Python project"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "langchain>=0.2.0,<0.3.0",
    "langchain-openai>=0.1.0,<0.2.0",
    "langchain-community>=0.2.0,<0.3.0",
    "langgraph>=0.2.0,<0.3.0",
    "langsmith>=0.1.0,<0.2.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "openai>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "ruff>=0.0.261",
    "mypy>=1.0",
    "pip-tools>=7.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true